+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Bugfixes:
 - fix problem of rendering an unspecified height table with no rows
 - `onBeforeChange` now allows to remove one change from array by assigning it null value (`changes[i] = null`)
 
+Other:
+- move `jquery.handsontable.js` and `jquery.handsontable.css` to `dist/`
+
 ## [0.8.23](https://github.com/warpech/jquery-handsontable/tree/v0.8.23) (May 3, 2013)
 
 Features: